// planning-idsdp-pps-frontend/src/app/main/modules/configuration-management/models/economic-code-model.ts
export class EconomicTypeModel {
    id: string;
    uuid: string;
    code: string;
    economicCodeFor: number;
    oldEconomicCodeEn: string;
    oldEconomicCodeBn: string;
    economicCode: string;
    economicCodeBng: string;
    economicCodeName: string;
    economicCodeNameBng: string;
    description: string;
    descriptionBn: string;
    status: boolean;
}